@@ -1,10 +1,10 @@
 module github.com/trezor/blockbook
 
-go 1.19
+go 1.21
 
 require (
-	github.com/ava-labs/avalanchego v1.9.11
-	github.com/ava-labs/coreth v0.11.8
+	github.com/ava-labs/avalanchego v1.10.18-rc.0
+	github.com/ava-labs/coreth v0.12.10-rc.0
 	github.com/bsm/go-vlq v0.0.0-20150828105119-ec6e8d4f5f4e
 	github.com/deckarep/golang-set v1.8.0
 	github.com/decred/dcrd/chaincfg/chainhash v1.0.2
