@@ -32,7 +32,11 @@ export interface EthereumSpecific {
     nonce: number;
     gasLimit: number;
     gasUsed?: number;
-    gasPrice: string;
+    gasPrice?: string;
+    l1Fee?: number;
+    l1FeeScalar?: string;
+    l1GasPrice?: string;
+    l1GasUsed?: number;
     data?: string;
     parsedData?: EthereumParsedInputData;
     internalTransfers?: EthereumInternalTransfer[];
@@ -46,9 +50,9 @@ export interface TokenTransfer {
     from: string;
     to: string;
     contract: string;
-    name: string;
-    symbol: string;
-    decimals: number;
+    name?: string;
+    symbol?: string;
+    decimals?: number;
     value?: string;
     multiTokenValues?: MultiTokenValue[];
 }
@@ -109,6 +113,17 @@ export interface FeeStats {
     averageFeePerKb: number;
     decilesFeePerKb: number[];
 }
+export interface StakingPool {
+    contract: string;
+    name: string;
+    pendingBalance: string;
+    pendingDepositedBalance: string;
+    depositedBalance: string;
+    withdrawTotalAmount: string;
+    claimableAmount: string;
+    restakedReward: string;
+    autocompoundBalance: string;
+}
 export interface ContractInfo {
     type: string;
     contract: string;
@@ -161,6 +176,7 @@ export interface Address {
     contractInfo?: ContractInfo;
     erc20Contract?: ContractInfo;
     addressAliases?: { [key: string]: AddressAlias };
+    stakingPools?: StakingPool[];
 }
 export interface Utxo {
     txid: string;
@@ -245,6 +261,7 @@ export interface InternalStateColumn {
 }
 export interface BlockbookInfo {
     coin: string;
+    network: string;
     host: string;
     version: string;
     gitCommit: string;
@@ -264,6 +281,7 @@ export interface BlockbookInfo {
     currentFiatRatesTime?: string;
     historicalFiatRatesTime?: string;
     historicalTokenFiatRatesTime?: string;
+    supportedStakingPools?: string[];
     dbSizeFromColumns?: number;
     dbColumns?: InternalStateColumn[];
     about: string;
@@ -338,6 +356,7 @@ export interface WsBackendInfo {
 export interface WsInfoRes {
     name: string;
     shortcut: string;
+    network: string;
     decimals: number;
     version: string;
     bestHeight: number;
@@ -356,6 +375,17 @@ export interface WsBlockReq {
     id: string;
     pageSize?: number;
     page?: number;
+}
+export interface WsBlockFilterReq {
+    scriptType: string;
+    blockHash: string;
+    M?: number;
+}
+export interface WsBlockFiltersBatchReq {
+    scriptType: string;
+    bestKnownBlockHash: string;
+    pageSize?: number;
+    M?: number;
 }
 export interface WsAccountUtxoReq {
     descriptor: string;
@@ -416,7 +446,17 @@ export interface WsFiatRatesTickersListReq {
 export interface WsMempoolFiltersReq {
     scriptType: string;
     fromTimestamp: number;
+    M?: number;
+}
+export interface WsRpcCallReq {
+    from?: string;
+    to: string;
+    data: string;
+}
+export interface WsRpcCallRes {
+    data: string;
 }
 export interface MempoolTxidFilterEntries {
     entries?: { [key: string]: string };
+    usedZeroedKey?: boolean;
 }
